Polish translation improvements
Mostly grammar updates, making it more clear and look more official, by using less colloquial words.
In some places I made it more understandable and true to original.
The most striking thing in previous translations is "Free" translated as "Free to download" not as "Free space" meaning.
I'm not sure if I should change the last translator on top, but I updated the revision date and filled the translators as they are now on the bottom.